Add new routes

+++ 02-Advanced/ExpressJS/01-RESTFul-Server/src/utility/file.js
@@ -1,11 +1,16 @@
 const fs = require("fs");
 
 const deleteFile = (filePath) => {
-  fs.unlink(filePath, (error) => {
-    if (error) {
-      throw new Error(error);
-    }
-  });
+  if (fs.existsSync(filePath)) {
+    fs.unlink(filePath, (error) => {
+      if (error) {
+        throw error;
+      }
+    });
+  } else {
+    // throw new Error("No such a file exists!");
+    console.error("No such a file exists!");
+  }
 };
 
 module.exports = { deleteFile };
